Refactor view folder structure according to Q-Tool-2021 state
