+++ views/components/Legend.svelte
