+++ views/components/PollResult.svelte
